+++ example/example-writer.py
@@ -1,5 +1,6 @@
 import random
 import copclib as copc
+import math
 
 random.seed(0)
 
@@ -75,8 +76,8 @@ def BoundsTrimFileExample():
     reader = copc.FileReader("autzen-classified.copc.laz")
     old_header = reader.las_header
 
-    # Take horizontal 2D box of [200,200] roughly in the middle of the point cloud.
-    box = copc.Box(637190, 851109, 637390, 851309)
+    middle = (old_header.max + old_header.min) / 2
+    box = copc.Box(middle.x - 200, middle.y - 200, middle.x + 200, middle.y + 200)
 
     # Copy the header to the new file
     cfg = reader.GetCopcConfig()
@@ -200,26 +201,27 @@ def RandomPoints(key, las_header, number_points):
         # Create a point with a given point format
         point = points.CreatePoint()
         # point has getters/setters for all attributes
-        point.UnscaledX = int(
-            random.uniform(
-                min(minx, las_header.max.x), min(minx + step, las_header.max.x)
-            )
+        point.UnscaledX = random.randint(
+            math.ceil(las_header.ApplyInverseScaleX(max(las_header.min.x, minx))),
+            math.floor(
+                las_header.ApplyInverseScaleX(min(las_header.max.x, minx + step))
+            ),
         )
-
-        point.UnscaledY = int(
-            random.uniform(
-                min(miny, las_header.max.y), min(miny + step, las_header.max.y)
-            )
+        point.UnscaledY = random.randint(
+            math.ceil(las_header.ApplyInverseScaleY(max(las_header.min.y, miny))),
+            math.floor(
+                las_header.ApplyInverseScaleY(min(las_header.max.y, miny + step))
+            ),
         )
-
-        point.UnscaledZ = int(
-            random.uniform(
-                min(minz, las_header.max.z), min(minz + step, las_header.max.z)
-            )
+        point.UnscaledZ = random.randint(
+            math.ceil(las_header.ApplyInverseScaleZ(max(las_header.min.z, minz))),
+            math.floor(
+                las_header.ApplyInverseScaleZ(min(las_header.max.z, minz + step))
+            ),
         )
 
         # For visualization purposes
-        point.PointSourceID = key.d + key.x + key.y + key.d
+        point.PointSourceID = key.d + key.x + key.y + key.z
 
         points.AddPoint(point)
     return points
@@ -229,28 +231,25 @@ def RandomPoints(key, las_header, number_points):
 def NewFileExample():
 
     # Create our new file with the specified format, scale, and offset
-    cfg = copc.CopcConfig(8, [1, 1, 1], [0, 0, 0])
+    cfg = copc.CopcConfig(8, [0.1, 0.1, 0.1], [50, 50, 50])
     # As of now, the library will not automatically compute the min/max of added points
     # so we will have to calculate it ourselves
-    cfg.las_header_base.min = (
-        MIN_BOUNDS * cfg.las_header_base.scale - cfg.las_header_base.offset
-    )
-    cfg.las_header_base.max = (
-        MAX_BOUNDS * cfg.las_header_base.scale - cfg.las_header_base.offset
-    )
+    cfg.las_header_base.min = MIN_BOUNDS
+    cfg.las_header_base.max = MAX_BOUNDS
+
     cfg.copc_info.spacing = 10
     cfg.wkt = "TEST_WKT"
 
     # Now, we can create our COPC writer:
     writer = copc.FileWriter("new-copc.copc.laz", cfg)
+    header = writer.las_header
 
     # The root page is automatically created
     root_page = writer.GetRootPage()
-    las_header = writer.las_header
 
     # First we'll add a root node
     key = copc.VoxelKey(0, 0, 0, 0)
-    points = RandomPoints(key, las_header, NUM_POINTS)
+    points = RandomPoints(key, header, NUM_POINTS)
     # The node will be written to the file when we call AddNode
     writer.AddNode(root_page, key, points)
 
@@ -261,20 +260,24 @@ def NewFileExample():
 
     # Once our page is created, we can add nodes to it like before
     key = copc.VoxelKey(1, 1, 1, 0)
-    points = RandomPoints(key, las_header, NUM_POINTS)
+    points = RandomPoints(key, header, NUM_POINTS)
     writer.AddNode(page, key, points)
 
     key = copc.VoxelKey(2, 2, 2, 0)
-    points = RandomPoints(key, las_header, NUM_POINTS)
+    points = RandomPoints(key, header, NUM_POINTS)
     writer.AddNode(page, key, points)
 
     # We can nest subpages as much as we want, as long as they are children of the parent
-    sub_page = writer.AddSubPage(page, (3, 4, 4, 2))
-    points = RandomPoints(sub_page.key, las_header, NUM_POINTS)
+    sub_page = writer.AddSubPage(page, copc.VoxelKey(3, 4, 4, 0))
+    points = RandomPoints(sub_page.key, header, NUM_POINTS)
     writer.AddNode(page, sub_page.key, points)
 
     # Make sure we call close to finish writing the file!
     writer.Close()
+
+    # We can check that the spatial bounds of the file have been respected
+    reader = copc.FileReader("new-copc.copc.laz")
+    assert reader.ValidateSpatialBounds()
 
 
 if __name__ == "__main__":