+++ test/reader_test.cpp
@@ -222,7 +222,7 @@ TEST_CASE("Spatial Query Functions", "[Reader]")
 {
     FileReader reader("autzen-classified.copc.laz");
 
-    // Make horizontal 2D box of [200,200] roughly in the middle of the point cloud.
+    // Make horizontal 2D box of [400,400] roughly in the middle of the point cloud.
     auto middle = (reader.GetLasHeader().max + reader.GetLasHeader().min) / 2;
     Box middle_box(middle.x - 200, middle.y - 200, middle.x + 200, middle.y + 200);